+++ nym-vpn-core/crates/nym-vpn-proto/src/conversions/into_proto/mod.rs
@@ -4,11 +4,11 @@
 pub mod account;
 pub mod account_shared_state;
 pub mod network_config;
+pub mod service;
 pub mod tunnel_event;
 pub mod tunnel_state;
 pub mod vpn_api_client;
 pub mod vpnd;
-pub mod service;
 
 impl From<String> for crate::Url {
     fn from(url: String) -> Self {
