+++ nym-vpn-core/crates/nym-vpn-proto/src/conversions/into_proto/service.rs
@@ -1,11 +1,11 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use nym_vpn_lib_types::VpnServiceInfo;
+use nym_vpn_lib_types::{VpnServiceConnectError, VpnServiceInfo};
 
 use crate::{
-    InfoResponse as ProtoInfoResponse, NymNetworkDetails as ProtoNymNetworkDetails,
-    NymVpnNetworkDetails as ProtoNymVpnNetworkDetails,
+    ConnectRequestError as ProtoConnectRequestError, InfoResponse as ProtoInfoResponse,
+    NymNetworkDetails as ProtoNymNetworkDetails, NymVpnNetworkDetails as ProtoNymVpnNetworkDetails,
 };
 
 impl From<VpnServiceInfo> for ProtoInfoResponse {
@@ -25,6 +25,21 @@ impl From<VpnServiceInfo> for ProtoInfoResponse {
             git_commit: info.git_commit,
             nym_network,
             nym_vpn_network,
+        }
+    }
+}
+
+impl From<VpnServiceConnectError> for ProtoConnectRequestError {
+    fn from(err: VpnServiceConnectError) -> Self {
+        match err {
+            VpnServiceConnectError::Internal(ref _account_error) => ProtoConnectRequestError {
+                kind: crate::connect_request_error::ConnectRequestErrorType::Internal as i32,
+                message: err.to_string(),
+            },
+            VpnServiceConnectError::Cancel => ProtoConnectRequestError {
+                kind: crate::connect_request_error::ConnectRequestErrorType::Internal as i32,
+                message: err.to_string(),
+            },
         }
     }
 }
