+++ nym-vpn-core/crates/nym-vpnd/src/command_interface/protobuf/error.rs
@@ -1,26 +1,7 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use crate::service::{SetNetworkError, VpnServiceConnectError, VpnServiceDeleteLogFileError};
-
-impl From<VpnServiceConnectError> for nym_vpn_proto::ConnectRequestError {
-    fn from(err: VpnServiceConnectError) -> Self {
-        match err {
-            VpnServiceConnectError::Internal(ref _account_error) => {
-                nym_vpn_proto::ConnectRequestError {
-                    kind: nym_vpn_proto::connect_request_error::ConnectRequestErrorType::Internal
-                        as i32,
-                    message: err.to_string(),
-                }
-            }
-            VpnServiceConnectError::Cancel => nym_vpn_proto::ConnectRequestError {
-                kind: nym_vpn_proto::connect_request_error::ConnectRequestErrorType::Internal
-                    as i32,
-                message: err.to_string(),
-            },
-        }
-    }
-}
+use crate::service::{SetNetworkError, VpnServiceDeleteLogFileError};
 
 #[derive(Debug, thiserror::Error)]
 pub enum VpnCommandSendError {
