+++ nym-vpn-core/crates/nym-vpnd/src/service/error.rs
@@ -7,23 +7,6 @@ use tracing::error;
 
 use super::config::ConfigSetupError;
 
-// Failure to initiate the connect
-#[derive(Clone, Debug, thiserror::Error)]
-pub enum VpnServiceConnectError {
-    #[error("internal error: {0}")]
-    Internal(String),
-
-    #[error("connection attempt cancelled")]
-    Cancel,
-}
-
-// Failure to initiate the disconnect
-#[derive(Clone, Debug, thiserror::Error)]
-pub enum VpnServiceDisconnectError {
-    #[error("internal error: {0}")]
-    Internal(String),
-}
-
 #[derive(Debug, thiserror::Error)]
 pub enum AccountError {
     // Failures related to the operational aspects of the account controller
