+++ nym-vpn-core/crates/nym-vpnd/src/service/mod.rs
@@ -11,10 +11,5 @@ pub use config::{
     config_dir, create_config_file, log_dir, read_config_file, write_config_file,
     DEFAULT_GLOBAL_CONFIG_FILE, DEFAULT_LOG_FILE,
 };
-pub use error::{
-    AccountError, SetNetworkError, VpnServiceConnectError, VpnServiceDeleteLogFileError,
-    VpnServiceDisconnectError,
-};
-pub use vpn_service::{
-    ConnectArgs, ConnectOptions, NymVpnService, VpnServiceCommand,
-};
+pub use error::{AccountError, SetNetworkError, VpnServiceDeleteLogFileError};
+pub use vpn_service::{ConnectArgs, ConnectOptions, NymVpnService, VpnServiceCommand};
